render avatar in group message & group participants sidebar
