+++ src/utils/styles/index.tsx
@@ -274,7 +274,7 @@ export const MessageItemContainer = styled.div`
   word-break: break-word;
 `;
 
-export const MessageItemAvatarStyle = styled.img`
+export const UserAvatarContainer = styled.img`
   width: 50px;
   height: 50px;
   border-radius: 50%;
